address pr 150: the shared process_* callback helper now lives in mpesa client and both client inherit it.Test asserts that both clients are instances of MpesaCallbackMixin
