+++ stubs/mpesakit/mpesa_client.pyi
@@ -1,7 +1,6 @@
 from _typeshed import Incomplete
 from mpesakit.auth import TokenManager as TokenManager
 from mpesakit.auth import AsyncTokenManager as AsyncTokenManager
-from mpesakit.callback_mixin import MpesaCallbackMixin as MpesaCallbackMixin
 from mpesakit.http_client import MpesaHttpClient as MpesaHttpClient
 from mpesakit.http_client import MpesaAsyncHttpClient as MpesaAsyncHttpClient
 from mpesakit.services import (
@@ -28,6 +27,20 @@ from mpesakit.services import (
     TransactionService as TransactionService,
     AsyncTransactionService as AsyncTransactionService,
 )
+
+class MpesaCallbackMixin:
+    def process_stk_callback(self, payload): ...
+    def process_stk_query_callback(self, payload): ...
+    def process_account_balance_callback(self, payload): ...
+    def process_account_balance_timeout(self, payload): ...
+    def process_b2c_callback(self, payload): ...
+    def process_b2b_callback(self, payload): ...
+    def process_transactions_callback(self, payload): ...
+    def process_bill_manager_callback(self, payload): ...
+    def process_tax_remittance_callback(self, payload): ...
+    def process_dynamic_qr_code_callback(self, payload): ...
+    def process_ratiba_service_callback(self, payload): ...
+    def process_reversal_callback(self, payload): ...
 
 class MpesaClient(MpesaCallbackMixin):
     http_client: Incomplete
